Move the cookie parser code to another file.
